Gra Quiz

+++ src/Model/Models/Database/GameSessionTranslation.cs
@@ -17,7 +17,7 @@ namespace Model.Models.Database
         [Required]
         public virtual bool Correct { get; set; }
 
-        public virtual int Duration { get; set; }
+        public virtual decimal Duration { get; set; }
 
         [NotMapped]
         public override bool IsValid => GameSessionId > 0 && TranslationId > 0;
